v0.5.1: - Move SlashCommandClient and DefaultSlashCommandClient from package. - Move SlashUtils.kt from package. - Add DiscoveryResult.kt to SlashUtils#discoverSlashCommands(). - Move internal permission checks to Checks.kt.
