+++ src/main/kotlin/tv/blademaker/slash/api/DiscoveryResult.kt
@@ -1,0 +1,7 @@
+package tv.blademaker.slash.api
+
+data class DiscoveryResult(
+    val elapsedTime: Long,
+    val count: Int,
+    val commands: List<BaseSlashCommand>
+)
