+++ src/main/kotlin/tv/blademaker/slash/api/SlashCommandContextImpl.kt
@@ -1,6 +1,7 @@
 package tv.blademaker.slash.api
 
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent
+import tv.blademaker.slash.client.SlashCommandClient
 import java.util.concurrent.atomic.AtomicReference
 
 class SlashCommandContextImpl(override val commandClient: SlashCommandClient,
