+++ src/main/kotlin/tv/blademaker/slash/extensions/JdaUtils.kt
@@ -4,7 +4,7 @@ import net.dv8tion.jda.api.JDA
 import net.dv8tion.jda.api.JDABuilder
 import net.dv8tion.jda.api.sharding.DefaultShardManagerBuilder
 import net.dv8tion.jda.api.sharding.ShardManager
-import tv.blademaker.slash.api.SlashCommandClient
+import tv.blademaker.slash.client.SlashCommandClient
 
 fun JDABuilder.setSlashCommandClient(slashCommandClient: SlashCommandClient) {
     addEventListeners(slashCommandClient)
